Change hashed-string input type from text to hidden
The hashed-string is a concatenation of all key-value pairs which are
used to calculate the content hash (fingerprint). It is added to the
form and used by the server for additional integrity checks. Because
one of the fields (Weitere Anmerkungen) is a textarea, it could
contain newlines, and hence the hashed-string could also contain
newlines. These would get stripped when entered (via JavaScript) into
the text field, causing integrity check failures on the
server. Apparently, this does not happen for input type=hidden (which
also does not need the style parameter to make it hidden).

https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/input/hidden

Fixes #44