Add HTML component for new go-no-go
The old go-no-go is renamed (just the header) to Go/Nogo-Task Niessen,
an a new assesment, Untertest Go/NoGo – Variante: 2 aus 5, is added.

This introduces a new kind of input field validation: percentile rank
(Prozentrang): 1-100 with an optional "<" or ">". Out of necessity, the
field is a string field, and the validation is done with a regex. To
reduce the variability of inputs, we enforce the </> symbol to be
followed by space.

This change concerns just the html (plus required toggle), not hashing
or saving the data.